WiningMovecalculator and BlockingWiningMove Done
CalculateWiningMove DONE
CalculateBlocke DONE

Is imposible to beat the program, he will block any wining move that you can try and will create his own wining move if you give him a chance.
for now this is ONLY BACKEND, I did a little interface on terminal where it can be tested and play.

What is left?

Make the engine recognize if he or the game wins and end up the game.